VB Shopping Cart Project
DISCLAIMER: for this project our class was given a skeleton of an app
with a main menu and we had to implement a general login and add
functionality for a specific feature; in my case, the shopping Cart
("Shop" section).

This program authenticates users and allows them to add different
amounts of searchable items to a cart. After completing a transaction,
the users get a report of their purchase that is then saved in the
database.

The SQL database used in this project contains authorized users,
products, customers, and invoices among other things.